+++ ydb/tests/stress/kv_volume/__main__.py
@@ -3,7 +3,7 @@ import argparse
 import logging
 import sys
 from google.protobuf import json_format, text_format
-from ydb.tests.stress.kv_volume.workload import YdbKeyValueVolumeWorkload
+from ydb.tests.stress.kv_volume.workload import YdbKeyValueVolumeWorkload, generate_preset_configs
 import ydb.tests.stress.kv_volume.protos.config_pb2 as config_pb
 
 
@@ -40,6 +40,8 @@ def load_config(path):
 
 
 if __name__ == '__main__':
+    generated_configs = generate_preset_configs()
+
     parser = argparse.ArgumentParser(
         description="Workload kv wrapper", formatter_class=argparse.RawDescriptionHelpFormatter
     )
@@ -49,7 +51,10 @@ if __name__ == '__main__':
     parser.add_argument('--in-flight', default=1, type=int, help='In flight')
     parser.add_argument('--log-file', default=None, help='Append log into specified file')
     parser.add_argument('--version', default='v1', choices=['v1', 'v2'], help='Keyvalue grpc api version')
-    parser.add_argument('--config', required=True, help='Path to KeyValueVolumeStressLoad config (textproto/json/binary)')
+
+    config_group = parser.add_mutually_exclusive_group(required=True)
+    config_group.add_argument('--config', help='Path to KeyValueVolumeStressLoad config (textproto/json/binary)')
+    config_group.add_argument('--config-name', choices=generated_configs.keys(), help='Name of preset config to use')
 
     args = parser.parse_args()
 
@@ -64,7 +69,14 @@ if __name__ == '__main__':
 
     # Resolve parameters, optionally overriding from protobuf config
     cfg = None
-    if args.config:
+    if args.config_name:
+        preset_configs = generate_preset_configs()
+        if args.config_name not in preset_configs:
+            logging.error("Unknown preset config name: %s", args.config_name)
+            logging.info("Available presets: %s", ', '.join(preset_configs.keys()))
+            sys.exit(1)
+        cfg = preset_configs[args.config_name]
+    else:
         cfg = load_config(args.config)
 
     workload = YdbKeyValueVolumeWorkload(
